pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inferx"
version = "1.0.0"
description = "Lightweight ML Inference Runtime"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "InferX Team", email = "team@inferx.dev"}
]
requires-python = ">=3.11"

dependencies = [
    "onnxruntime>=1.16.0",
    "openvino>=2023.3.0",
    "click>=8.1.0",
    "pyyaml>=6.0.1",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.16.0",
]

serve = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
inferx = "inferx.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["inferx"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]